feat: Refactor GraphQL module structure and implement user object with queries and mutations
