+++ apps/backend/src/presentation/graphql/objects/user_object.rs
